# docs/swagger.yaml
basePath: /api
definitions:
  domain.Seller:
    properties:
      created_at:
        type: string
      description:
        type: string
      store_name:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handler.CreateSellerRequest:
    properties:
      description:
        example: We sell the coolest gadgets around.
        type: string
      email:
        example: seller@example.com
        type: string
      password:
        example: seller123
        type: string
      roles:
        example:
        - seller
        items:
          type: string
        type: array
      store_name:
        example: Cool Gadgets Store
        type: string
    type: object
  handler.CreateShipmentRequest:
    properties:
      order_id:
        example: ord-123456
        type: string
      provider:
        example: Kerry Express
        type: string
      tracking_number:
        example: KER123456789TH
        type: string
    type: object
...
  handler.UpdateShipmentStatusRequest:
    properties:
      status:
        example: in_transit
        type: string
    type: object
  handler.UpdateShipmentTrackingRequest:
    properties:
      provider:
        example: Kerry Express
        type: string
      tracking_number:
        example: KER987654321TH
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample e-commerce server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Golang E-Commerce API
  version: "1.0"
paths:
...
  /admin/sellers:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all sellers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all sellers (Admin)
      tags:
      - admin
      - sellers
    post:
      consumes:
      - application/json
      description: Create a new seller account and store
      parameters:
      - description: Seller details (email, password, roles, store_name, description)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSellerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new seller (Admin)
      tags:
      - admin
      - sellers
  /admin/sellers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a seller account
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a seller (Admin)
      tags:
      - admin
      - sellers
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get seller by ID (Admin)
      tags:
      - admin
      - sellers
    put:
      consumes:
      - application/json
      description: Update details of an existing seller
      parameters:
      - description: Seller ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated seller details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Seller'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a seller (Admin)
      tags:
      - admin
      - sellers
  /admin/shipments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Update the status of a shipment
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Status update (status)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateShipmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update shipment status (Admin)
      tags:
      - admin
      - shipments
  /admin/shipments/{id}/tracking:
    patch:
      consumes:
      - application/json
      description: Update the tracking number and provider for a shipment
      parameters:
      - description: Shipment ID
        in: path
        name: id
        required: true
        type: string
      - description: Tracking info (provider, tracking_number)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateShipmentTrackingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update shipment tracking (Admin)
      tags:
      - admin
      - shipments
  /admin/shipments/create:
    post:
      consumes:
      - application/json
      description: Create a shipment for an order
      parameters:
      - description: Shipment details (order_id, provider, tracking_number)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateShipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a shipment (Admin)
      tags:
      - admin
      - shipments
...
  /user/shipments/order/{orderId}:
    get:
      consumes:
      - application/json
      description: Retrieve shipment information for a specific order
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get shipment by order
      tags:
      - shipments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"